Refactoring: move file reading code from Template to util
