+++ engine/testdata/readfile-plain.txt
@@ -1,0 +1,1 @@
+foobar
